requirements.txt
Flask==2.3.2
flasgger==0.9.5
rembg==2.0.32
Pillow==9.3.0
